Gym update, gym class divided to gym and mlm gym
