+++ .gitignore
@@ -2,6 +2,8 @@
 # Custom Dzeniks
 *.pth
 *.bin
+*.json
+*.jsonl
 track.jsonl
 
 # Byte-compiled / optimized / DLL files
