+++ backend/gym.py
@@ -1,10 +1,8 @@
-import time
-from time import strftime
 import torch
 import json
 import os
 
-class Gym_albert():
+class Gym:
     def __init__(self, model, tokenizer, name="Model", epochs = 0, loss = 0) -> None:
         self.name = name
         self.model = model
@@ -13,33 +11,6 @@ class Gym_albert():
         self.track = []
         self.modelPath = f"{name}.pth"
         self.loss = loss
-
-    def train(self, data_loader, loss_fn, optimizer) -> None:
-        data_size = len(data_loader.dataset)
-        train_track = {"epoch": len(self.track), "loss": [], "batch_size" : data_loader.batch_size, "dataset_size": data_size}
-        self.model.train()
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-        for batch, (x, y) in enumerate(data_loader):
-
-            inputIDs = x["input_ids"].squeeze(1).to(device)
-            attention = x["attention_mask"].squeeze(1).to(device)
-            typeIds = x["token_type_ids"].squeeze(1).to(device)
-
-            y = y.to(device)
-
-            prediction = self.model(
-                input_ids=inputIDs, attention_mask=attention, token_type_ids=typeIds)
-            loss = loss_fn(prediction.logits, y)
-
-            optimizer.zero_grad()
-            loss.backward()
-            optimizer.step()
-            if batch % 100 == 0:
-                loss, progress = loss.item(), batch * len(y)
-                log = f"loss: {loss:>8f}, [{progress:>5f}/{data_size:>5f}]\n"
-                train_track["loss"].append(loss)
-                print(log)
-        self.track.append(train_track)
 
     def calculate_acc_loss_avg(self, corrects, loss, batch_num, lossFn, prediction, label):
         loss += lossFn(prediction.logits, label)
@@ -55,63 +26,6 @@ class Gym_albert():
         loss += lossFn(prediction.logits, label)
         return corrects, loss
 
-    def test(self, data_loader, loss_fn) -> None:
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-        dataSize = len(data_loader.dataset)
-        numBatches = len(data_loader)
-        self.model.eval()
-        loss, correct, correct2, loss2 = 0, 0, 0, 0
-        with torch.no_grad():
-            for batch, (x, y) in enumerate(data_loader):
-
-                inputIDs = x["input_ids"].squeeze(1).to(device)
-                attention = x["attention_mask"].squeeze(1).to(device)
-                typeIds = x["token_type_ids"].squeeze(1).to(device)
-
-                y = y.to(device)
-
-                prediction = self.model(
-                    input_ids=inputIDs, attention_mask=attention, token_type_ids=typeIds)
-
-                correct, loss = self.calculate_acc_loss_avg(
-                    correct, loss, batch, loss_fn, prediction, y)
-                correct2, loss2 = self.calculate_acc_loss_press(
-                    correct2, loss2, loss_fn, prediction, y, batch)
-
-        loss /= numBatches
-        loss2 /= numBatches
-        correct /= dataSize
-        correct2 /= dataSize
-        self.loss = loss
-        self.acc = correct
-        log = f"Results: \n Test Error: {(100*correct):>0.1f}%, Avg loss: {loss:>8f} \n My Accuracy: {(100*correct2):>0.1f}%, Avg loss: {loss2:>8f}\n"
-        print(log)
-
-    def test_sqce(self, data_loaders, lossFn):
-        start_time = time.perf_counter()
-        for i, data_loader in enumerate(data_loaders):
-            EpochLog = f"\nTest {i + 1} \n---------------------------------\n"
-            print(EpochLog)
-            self.test(data_loader, lossFn)
-            epochTime = f"--- {(time.perf_counter() - start_time)} seconds ---\n"
-            print(epochTime)
-            end = strftime(f"Test Ended!!! %H%M-%d-%m")
-            print(end)
-
-    def train_sqce(self, epochs, train_data_loader, test_data_loaders, lossFn, optimizer) -> None:
-        start_time = time.perf_counter()
-        for i in range(epochs):
-            EpochLog = f"\nEpoch {i + 1} \n---------------------------------\n"
-            print(EpochLog)
-            self.train(train_data_loader, lossFn, optimizer)
-            epochTime = f"--- {(time.perf_counter() - start_time)} seconds ---\n"
-            print(epochTime)
-            torch.save(self.model.state_dict(), "LastBackUp.pth")
-        self.epochs += epochs
-        self.save_track()
-        self.save_model()
-        self.test_sqce(test_data_loaders, lossFn)
-
     def save_track(self):
         n = 0
         filename = f"track{n}.jsonl"
@@ -124,7 +38,6 @@ class Gym_albert():
             for i in self.track:
                 json.dump(i, f)
                 f.write("\n")
-
 
     def save_model(self) -> None:
         torch.save(
